Update xmp parsing logic
