+++ .github/workflows/lint_and_test.yaml
@@ -26,7 +26,7 @@ jobs:
           version: 1.6.1
 
       - name: Install Dependencies
-        run: poetry install -E timestamps
+        run: poetry install --all-extras
 
       - name: Pre-Commit
         run: poetry run pre-commit run --all-files
